==> 07-OneHotEncoding.ipynb <==
***Dr. Emmanuel Dufourq*** www.emmanueldufourq.com

***African Institute for Mathematical Sciences***

***Stellenbosch University***

***2019***


## Download Keras

devtools::install_github("rstudio/keras")

## Import the Keras library

library(keras)

## Example data

example1=c(2, 1, 2, 0)

## Conver to categorical equivalent

to_categorical(example1, 3)

## Example data 2

example2=c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9)

## Convert to categorical equivalent

to_categorical(example2, 10)
