==> 07-Softmax.ipynb <==
***Dr. Emmanuel Dufourq*** www.emmanueldufourq.com

***African Institute for Mathematical Sciences***

***Stellenbosch University***

***2019***


## Download Keras

devtools::install_github("rstudio/keras")

## Import the Keras library

library(keras)

## Code to compute softmax function

logsumexp <- function (x) {
  y = max(x)
  y + log(sum(exp(x - y)))
}

softmax <- function (x) {
  exp(x - logsumexp(x))
}

## Example data

example1=c(2, 1, 2, 0, 3)

## Compute softmax

softmax(example1)

## Check that it sums to one

sum(softmax(example1))
