==> 12-EmbeddingRecommendationEngine.ipynb <==
***Dr. Emmanuel Dufourq*** www.emmanueldufourq.com

***African Institute for Mathematical Sciences***

***2019***


Credits: This notebook is an extension of the one provided by Prof. Ian Durbach (https://github.com/iandurbach/) link to notebook: https://github.com/iandurbach/datasci-fi/blob/921f19903c20db3b3a957da5ec80d78988a29376/lesson6-more-nns.ipynb

***NOTE***

Be sure to use hardware acceleration to use the GPU. Click on `Runtime`, change `runtime type`, and select `GPU` for the *hardware accelerator* option.




In this example we build a recommender system for the full "small" MovieLens dataset. Previously we saw how to use matrix decomposition to represent each movie and each user as a vector of latent variables. Here we use neural networks to learn the "weights" in these latent factors.

devtools::install_github("rstudio/keras")

library(tidyverse)
library(keras)

### Load the data

load(url("https://github.com/iandurbach/datasci-fi/raw/master/data/movielens-small.RData"))

### Look at what is available

ls()

### Check the "ratings" data

ratings

### Subtract 1 from the user ID and movie ID

ratings <- ratings %>% mutate(userId = -1 + as.numeric(factor(userId)),
                              movieId = -1 + as.numeric(factor(movieId)))

ratings

### Set the number of users and movies

n_users <- length(unique(ratings$userId))
n_movies <- length(unique(ratings$movieId))

n_users

n_movies

### Choose the number of dimensions to use in each embedding

n_factors <- 50

### Split the data into training and testing

Randomly assign 80% of the ratings to the training data and keep the remaining 20% aside as test data.

train_indicator <- (runif(nrow(ratings)) < 0.8)
training_ratings <- ratings[train_indicator,]
test_ratings <- ratings[-train_indicator,]

### A note on the Keras Function API

The functional API enables one to create more complex models than the ones which can be created using Sequential models. For instance, the functional API allows one to create a model which multiple inputs and outputs.

With the functional API the main architecture is defined by the `keras_model` class. This takes in inputs and outputs. Further information here: https://keras.rstudio.com/reference/keras_model.html

We can then define the inputs and outputs separately and then bring everything together.

We will begin by defining two input layers and then one output layer, finally we bring all three of these together using `keras_model`.

Further information: https://keras.rstudio.com/articles/functional_api.html

### Define two inputs layers

Below we create two inputs layers, one for the user and one for the movie. These are created separately.

We specify the shape of our input layers for user and movie embeddings. These are just a single value, representing the index of the user or movie.

user_in <- layer_input(shape = c(1), dtype = 'int64', name = 'user_in')
movie_in <- layer_input(shape = c(1), dtype = 'int64', name = 'movie_in')

### Define two embedding layers (which connect to the two inputs)

Next we define two separate embeddings layers, one which takes the user input and the other takes the movie input.

The user embedding will be have dimension *n_users* x *n_factors*

The movie embedding will be have dimension *n_movies* x *n_factors*

In both cases there is only a single value input.

### Question: how many parameters will there be in the user and movie embedding layers? How would each embedding layer roughly look like if you had to draw a simple representation of it?


user_emb <- user_in %>% layer_embedding(input_dim = n_users, output_dim = n_factors, input_length = 1)
movie_emb <- movie_in %>% layer_embedding(input_dim = n_movies, output_dim = n_factors, input_length = 1)

### Define one output layer

We now "combine" the embeddings and then proceed as usual.

To combine the embeddings we simply concatenate the embeddings and then flatten that.

predictions <- layer_concatenate(c(user_emb, movie_emb)) %>%
  layer_flatten() %>% 
  layer_dropout(0.3) %>%
  layer_dense(70, activation='relu') %>% 
  layer_dropout(0.75) %>%
  layer_dense(1)

### Create the Model

Now that we have created the two inputs and one output, we can go ahead and create the final model which links everything together.

We added all the complexities to the input and outputs - so combining everything is simplier in terms of code.

model <- keras_model(c(user_in, movie_in), predictions)

### Print out a summary of the model

Notice how there are two inputs which are both followed by an embedding layer. Then notice how the embeddings are concatenated together (each had 50 dimensions which results in a final shape of 100).

summary(model)

### Question: Can you roughly draw what the network looks like given this information?

### Compile

Just like before, nothing new here

model %>% compile(optimizer='Adam', loss='mse')

### What data do we input into the NN?

In our network we specified that there are two inputs, one for the user id and the other for the movie id. Let's create a list which contains these two data arrays and check the dimension.

x_train = as.matrix(list(training_ratings$userId, training_ratings$movieId))
y_train = as.matrix(training_ratings$rating)
dim(x_train)
dim(y_train)

### Fitting the model

Very similar to before. We provide the training features, training targets and the number of epochs.

model %>% fit(x = x_train, 
           y = y_train, 
           batch_size=64, 
           epoch=2)

### Evaluate the performance on the test data

model %>% evaluate(list(test_ratings$userId, test_ratings$movieId), 
                test_ratings$rating)

### Predicting on the test data

model %>% predict(list(test_ratings$userId, test_ratings$movieId))
